Removing direct references to ChannelImpl that don't need it.
