+++ core/src/main/java/io/grpc/internal/Http2ClientStream.java
@@ -166,7 +166,7 @@ public abstract class Http2ClientStream extends AbstractClientStream<Integer> {
   private static Status statusFromHttpStatus(Metadata metadata) {
     Integer httpStatus = metadata.get(HTTP2_STATUS);
     if (httpStatus != null) {
-      Status status = HttpUtil.httpStatusToGrpcStatus(httpStatus);
+      Status status = GrpcUtil.httpStatusToGrpcStatus(httpStatus);
       return status.isOk() ? status
           : status.augmentDescription("extracted status from HTTP :status " + httpStatus);
     }
@@ -204,8 +204,8 @@ public abstract class Http2ClientStream extends AbstractClientStream<Integer> {
       return null;
     }
     contentTypeChecked = true;
-    String contentType = headers.get(HttpUtil.CONTENT_TYPE_KEY);
-    if (TEMP_CHECK_CONTENT_TYPE && !HttpUtil.CONTENT_TYPE_GRPC.equalsIgnoreCase(contentType)) {
+    String contentType = headers.get(GrpcUtil.CONTENT_TYPE_KEY);
+    if (TEMP_CHECK_CONTENT_TYPE && !GrpcUtil.CONTENT_TYPE_GRPC.equalsIgnoreCase(contentType)) {
       // Malformed content-type so report an error
       return Status.INTERNAL.withDescription("invalid content-type " + contentType);
     }
@@ -216,7 +216,7 @@ public abstract class Http2ClientStream extends AbstractClientStream<Integer> {
    * Inspect the raw metadata and figure out what charset is being used.
    */
   private static Charset extractCharset(Metadata headers) {
-    String contentType = headers.get(HttpUtil.CONTENT_TYPE_KEY);
+    String contentType = headers.get(GrpcUtil.CONTENT_TYPE_KEY);
     if (contentType != null) {
       String[] split = contentType.split("charset=");
       try {
